dev/project_initialization Refactored the auth middleware 2
