+++ src/user/interface/user.repository.interface.ts
@@ -1,4 +1,4 @@
-import { User } from './user.entity';
+import { User } from '../user.entity';
 import { UserModel } from '@prisma/client';
 
 export interface IUserRepository {
